ci: [android] better multi crate build
