+++ .gitignore
@@ -6,6 +6,8 @@
 **/**/.idea/*
 sifir-android/outputs/**
 sifir-ios/outputs/**
-sifir-ios/scripts/custom/**
-./sifir-android/src/ffi/java_glue.rs
-./sifir-android/outputs/aar/**
+sifir-ios/scripts/custom/**a
+# android builds
+sifir-android/src/ffi/java_glue.rs
+sifir-android/outputs/aar/**
+#sifir-android/app/tor/src/main/java/com/sifir/**
